# refactor(digest:sha3): optimizing sha3 implementation