inlining/unrolling message scheduling and compression function

SHA3 Performance:
* Intel Xeon E5-1650 v3 (Haswell):
  sha3-256: 108 MB/s => 299 MB/s
  sha3-512:  57 MB/s => 162 MB/s
* AMD A10-7850K (Kaveri):
  sha3-256: 87 MB/s => 206 MB/s
  sha3-512: 48 MB/s => 110 MB/s