@homu 2016-01-15
:sunny: Test successful - [status](https://travis-ci.org/libOctavo/octavo/builds/102667294)